Changed camera snapping + cleaned up files
I reimplemented the camera controls, but I didn't line it up perfectly so you will have to fiddle with that by moving around each area and/or changing the area2d collision. I deleted a bunch of unused files but feel free to bring it back if you are planning to use it. I also moved the position of some files.